requirements.txt
requests
bencode.py